requirements.txt
python-telegram-bot==20.8
openai>=1.6.0
python-dotenv>=1.0.0

